guanghu-self-hosted/boot-protocol/requirements.txt
# Boot Protocol 运行时依赖
# Phase-0-006 · YD-A05-20260425-004
# Python 3.11+

asyncpg>=0.29.0
PyYAML>=6.0
toml>=0.10.2
pydantic>=2.0
